+++ packages/utils/src/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './path-mapper'
+export * from './logger'